Get Current User From Authentication In Firebase
to compare the id in the UserPage url  to the Authentication User id, and not to the id in local storage, to prevent hacks.
all of the data that is presented in UserPage is from firebase directly, except the user address which is from userDetails in LocalStorage.
also, the user orders info is imported from firebase with the id of the true current user and not the user written in local storage.