Get All Products In Cart Page From Firebase
instead of getting products details from LocalStorage, to prevent hacking and changing details of products, like their prices.
when viewing products in Home page and in Search and in ProductsPage, the data is still taken from Local Storage, because there is no affect there on the firebase data and on the price of the orders, it's only in Cart page.